Added unit tests for random functions and fixed structure
