+++ src/mygrad/random/__init__.py
@@ -1,0 +1,1 @@
+from .random import *
